Use `i64` to accomodate for device files' major and minor numbers
This is meant as the final step in fixing ogham/exa#1126.

The different build targets of Rust libc's `major()` and `minor()`
functions have conflicting return types like `c_uint`, `c_int` and
`i32`, so an `i64` should be able to represent all of them.

+++ src/fs/fields.rs
@@ -194,8 +194,8 @@ pub enum Size {
 /// - <http://www.lanana.org/docs/device-list/devices-2.6+.txt>
 #[derive(Copy, Clone)]
 pub struct DeviceIDs {
-    pub major: u32,
-    pub minor: u32,
+    pub major: i64,
+    pub minor: i64,
 }
