+++ src/fs/file.rs
@@ -336,9 +336,11 @@ impl<'dir> File<'dir> {
         else if self.is_char_device() || self.is_block_device() {
             let device_ids = self.metadata.rdev();
 
+            // SAFETY: `major()` and `minor()` can return an unsigned integer
+            // of at most 32bits, so an `i64` is going to accomodate them well
             f::Size::DeviceIDs(f::DeviceIDs {
-                major: unsafe { major(device_ids) },
-                minor: unsafe { minor(device_ids) },
+                major: unsafe { major(device_ids) as i64 },
+                minor: unsafe { minor(device_ids) as i64 },
             })
         }
         else {
